+++ main/localization/settings.go
@@ -1,0 +1,27 @@
+package localization
+
+var (
+	// Categories
+	SettingCategoryFiles = Translations{
+		englishUS: "Files",
+	}
+	SettingCategoryDecentralization = Translations{
+		englishUS: "Decentralization",
+	}
+
+	// File settings
+	SettingFilesMaxUploadSize = Translations{
+		englishUS: "Maximum size of files that can be uploaded (in MB)",
+	}
+	SettingFilesMaxTotalStorage = Translations{
+		englishUS: "Maximum amount of storage one account can use (in MB)",
+	}
+
+	// Decentralization settings
+	SettingDecentralizationEnabled = Translations{
+		englishUS: "Enable decentralization",
+	}
+	SettingDecentralizationAllowUnsafe = Translations{
+		englishUS: "Allow unsafe connections (not recommended)",
+	}
+)