+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Removed unused routes (/app/\* and /node/manage/regen + remove)
 - Some code cleanup moving all entities into the database module
 - Account management for admins (rank changing, searching, deleting)
+- Moved file max size and total storage settings to the admin panel
+- New settings for decentralization in the admin panel